backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
pymongo==4.5.0
werkzeug==2.3.7
PyPDF2==3.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0
gunicorn==21.2.0 